Some tuning to filesystem operations

+++ src/file_storage.cpp
@@ -8,8 +8,12 @@ using namespace yappi::persistance::backends;
 namespace fs = boost::filesystem;
 
 file_storage_t::file_storage_t(const std::string& uuid, bool purge):
-    m_storage_path("/var/lib/yappi/" + uuid)
+    m_storage_path("/var/lib/yappi/" + uuid + ".tasks")
 {
+    if(fs::exists(m_storage_path) && !fs::is_directory(m_storage_path)) {
+        throw std::runtime_error(m_storage_path.string() + " is not a directory");
+    }
+
     if(purge) {
         syslog(LOG_NOTICE, "storage: purging");
         fs::remove_all(m_storage_path);
@@ -54,6 +58,7 @@ bool file_storage_t::exists(const std::string& key) const {
 Json::Value file_storage_t::get(const std::string& key) const {
     Json::Reader reader(Json::Features::strictMode());
     Json::Value root(Json::objectValue);
+
     fs::path filepath = m_storage_path / key;
     fs::ifstream stream;
      
@@ -66,7 +71,7 @@ Json::Value file_storage_t::get(const std::string& key) const {
     }
 
     if(!reader.parse(stream, root)) {
-        syslog(LOG_ERR, "storage: malformed object in %s - %s",
+        syslog(LOG_ERR, "storage: malformed json in %s - %s",
             filepath.string().c_str(), reader.getFormatedErrorMessages().c_str());
     }
